+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added support for more than 2 CLIP models in advanced mode.
 - Added authorization for HuggingFace downloads.
 - Added support for loading checkpoints that don't contain VAE/CLIP/UNET (e.g. SD 3.x) and then supplying external models.
+- Added a button to reset selection in model fields.
 
 ## Improvements